// src/lib/api/schemas/flowResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Onboarding-as-a-Service
 * Onboarding-as-a-Service (OaaS) Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { FlowResponseDescription } from './flowResponseDescription';
import type { FlowResponseDurationDays } from './flowResponseDurationDays';
import type { FlowResponseStatus } from './flowResponseStatus';

export interface FlowResponse {
  name: string;
  description?: FlowResponseDescription;
  duration_days?: FlowResponseDurationDays;
  status?: FlowResponseStatus;
  id: string;
  company_id: string;
  created_at: string;
  updated_at: string;
}