// lib/api/schemas/flowUpdate.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Onboarding-as-a-Service
 * Onboarding-as-a-Service (OaaS) Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { FlowUpdateName } from './flowUpdateName';
import type { FlowUpdateDescription } from './flowUpdateDescription';
import type { FlowUpdateDurationDays } from './flowUpdateDurationDays';
import type { FlowUpdateStatus } from './flowUpdateStatus';

export interface FlowUpdate {
  name?: FlowUpdateName;
  description?: FlowUpdateDescription;
  duration_days?: FlowUpdateDurationDays;
  status?: FlowUpdateStatus;
}